switched .h to .hpp, switched from pragma once to header guards, switched from classes to namespaces. started code for memory manager, all kinds of other stuff but thats the gist of it

+++ kernel/lib/memory.cpp
@@ -1,8 +1,8 @@
-#include "memory.h"
+#include "memory.hpp"
 
 extern "C" void* memcpy(void *destination, void *source, uint64_t size) {
 	uint8_t *destination_bytes = (uint8_t*) destination;
 	uint8_t *source_bytes = (uint8_t*) source;
-	for (int i = 0; i < size; i++) *(destination_bytes + i) = *(source_bytes + i);
+	for (uint64_t i = 0; i < size; i++) *(destination_bytes + i) = *(source_bytes + i);
 	return destination;
 }
